Add debug output back into the branch
Merge commit 92ac8b54c2a5cae5c3c771f66f6b9c1fa72206c0 was from release to this
branch, which forced inheritance of the removal of the debug output. This was
undesired and unforseen.
This commit partially reverts that merge by keeping the bugfix and adding back
the debug output statements.