Fix doorway bugs
Doorways CANNOT be stood on anymore. When entering a doorway cell from
its room, the game will try to push the player to the corridor on the other
side. If this is occupied, then the game declares the move invalid and aborts
it.
By the same token, entering any doorway cell from anything BUT its room results
in the game triggering a room entry on the doorway; pushing the player token
into the room and assigning them the next free space.